fix #58
