+++ src/main/java/moffy/ticex/modules/general/TicEXModule.java
@@ -123,7 +123,6 @@ public class TicEXModule extends AddonModule{
         bus.addListener(TicEXEvent::onEntityAttributeModification);
         bus.addListener(TicEXEvent::onRegisterCaps);
 
-        MinecraftForge.EVENT_BUS.addListener(TicEXEvent::onPlayerTick);
         MinecraftForge.EVENT_BUS.addListener(TicEXEvent::modifyAttribute);
         MinecraftForge.EVENT_BUS.addListener(EventPriority.LOW, TicEXEvent::onEntityHeal);
         MinecraftForge.EVENT_BUS.addListener(EventPriority.LOW, TicEXEvent::onEntityHurt);
